+++ src/base/tools/cryptonote/BlockTemplate.cpp
@@ -197,6 +197,11 @@ bool xmrig::BlockTemplate::parse(bool hashes)
         ar(m_vote);
     }
 
+    if (m_coin == Coin::ZEPHYR) {
+        uint8_t pricing_record[120];
+        ar(pricing_record);
+    }
+
     // Miner transaction begin
     // Prefix begin
     setOffset(MINER_TX_PREFIX_OFFSET, ar.index());
@@ -220,8 +225,12 @@ bool xmrig::BlockTemplate::parse(bool hashes)
     ar(m_height);
     ar(m_numOutputs);
 
-    // must be 1 output
-    if (m_numOutputs != 1) {
+    if (m_coin == Coin::ZEPHYR) {
+        if (m_numOutputs < 2) {
+            return false;
+        }
+    }
+    else if (m_numOutputs != 1) {
         return false;
     }
 
@@ -237,7 +246,37 @@ bool xmrig::BlockTemplate::parse(bool hashes)
 
     ar(m_ephPublicKey, kKeySize);
 
-    if (m_outputType == 3) {
+    if (m_coin == Coin::ZEPHYR) {
+        if (m_outputType != 2) {
+            return false;
+        }
+
+        uint64_t asset_type_len;
+        ar(asset_type_len);
+        ar.skip(asset_type_len);
+        ar(m_viewTag);
+
+        for (uint64_t k = 1; k < m_numOutputs; ++k) {
+            uint64_t amount2;
+            ar(amount2);
+
+            uint8_t output_type2;
+            ar(output_type2);
+            if (output_type2 != 2) {
+                return false;
+            }
+
+            Span key2;
+            ar(key2, kKeySize);
+
+            ar(asset_type_len);
+            ar.skip(asset_type_len);
+
+            uint8_t view_tag2;
+            ar(view_tag2);
+        }
+    }
+    else if (m_outputType == 3) {
         ar(m_viewTag);
     }
 
@@ -248,6 +287,8 @@ bool xmrig::BlockTemplate::parse(bool hashes)
     BlobReader<true> ar_extra(blob(TX_EXTRA_OFFSET), m_extraSize);
     ar.skip(m_extraSize);
 
+    bool pubkey_offset_first = true;
+
     while (ar_extra.index() < m_extraSize) {
         uint64_t extra_tag  = 0;
         uint64_t size       = 0;
@@ -256,7 +297,10 @@ bool xmrig::BlockTemplate::parse(bool hashes)
 
         switch (extra_tag) {
         case 0x01: // TX_EXTRA_TAG_PUBKEY
-            setOffset(TX_PUBKEY_OFFSET, offset(TX_EXTRA_OFFSET) + ar_extra.index());
+            if (pubkey_offset_first) {
+                pubkey_offset_first = false;
+                setOffset(TX_PUBKEY_OFFSET, offset(TX_EXTRA_OFFSET) + ar_extra.index());
+            }
             ar_extra.skip(kKeySize);
             break;
 
@@ -269,12 +313,19 @@ bool xmrig::BlockTemplate::parse(bool hashes)
         case 0x03: // TX_EXTRA_MERGE_MINING_TAG
             ar_extra(size);
             setOffset(TX_EXTRA_MERGE_MINING_TAG_OFFSET, offset(TX_EXTRA_OFFSET) + ar_extra.index());
-            ar_extra(m_txMergeMiningTag, size + kKeySize);
+            ar_extra(m_txMergeMiningTag, size);
             break;
 
         default:
             return false; // TODO(SChernykh): handle other tags
         }
+    }
+
+    if (m_coin == Coin::ZEPHYR) {
+        uint64_t pricing_record_height, amount_burnt, amount_minted;
+        ar(pricing_record_height);
+        ar(amount_burnt);
+        ar(amount_minted);
     }
 
     setOffset(MINER_TX_PREFIX_END_OFFSET, ar.index());